pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["DinarDi <dinktmt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"
django-debug-toolbar-force = "^0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
